# Search with equal priority in both z-order directions