This gives a slight speedup for more complex polygons.

Before:
```
typical OSM building (15 vertices): x 644,598 ops/sec ±0.96% (89 runs sampled)
dude shape (94 vertices): x 46,829 ops/sec ±1.39% (96 runs sampled)
dude shape with holes (104 vertices): x 37,816 ops/sec ±1.20% (91 runs sampled)
complex OSM water (2523 vertices): x 588 ops/sec ±0.75% (91 runs sampled)
```

After:
```
typical OSM building (15 vertices): x 632,252 ops/sec ±1.60% (89 runs sampled)
dude shape (94 vertices): x 48,956 ops/sec ±1.45% (94 runs sampled)
dude shape with holes (104 vertices): x 39,665 ops/sec ±1.51% (95 runs sampled)
complex OSM water (2523 vertices): x 659 ops/sec ±0.75% (91 runs sampled)
```

Results in gl-js:

https://bl.ocks.org/anonymous/raw/c415471d58758dd81650dec97d35b2f8/