@jfirebaugh 2018-01-03
I also tried searching in order of increasing z-distance from the current point -- i.e. advance only one of `p` and `n` in the main loop, whichever is currently closer in `z` to `ear.z`. But it benchmarked slower in all the variations that I tried.