@jfirebaugh 2018-01-03
Adding additional branches to the loop seems to negate the performance improvement. E.g.:

```js
    var p = ear.prevZ,
        n = ear.nextZ,
        hasPrev = p && p.z >= minZ,
        hasNext = n && n.z <= maxZ;

    // look for points inside the triangle in both directions
    while (hasPrev || hasNext) {
        if (hasPrev) {
            if (p !== ear.prev && p !== ear.next &&
                pointInTriangle(a.x, a.y, b.x, b.y, c.x, c.y, p.x, p.y) &&
                area(p.prev, p, p.next) >= 0) return false;
            p = p.prevZ;
            hasPrev = p && p.z >= minZ;
        }

        if (hasNext) {
            if (n !== ear.prev && n !== ear.next &&
                pointInTriangle(a.x, a.y, b.x, b.y, c.x, c.y, n.x, n.y) &&
                area(n.prev, n, n.next) >= 0) return false;
            n = n.nextZ;
            hasNext = n && n.z <= maxZ;
        }
    }
```

```
typical OSM building (15 vertices): x 606,847 ops/sec ±1.16% (87 runs sampled)
dude shape (94 vertices): x 29,195 ops/sec ±1.19% (91 runs sampled)
dude shape with holes (104 vertices): x 23,432 ops/sec ±1.06% (92 runs sampled)
complex OSM water (2523 vertices): x 359 ops/sec ±0.72% (87 runs sampled)
```